Removed implicit 'id' attribute on KML elements; added 'id' to constructors
This commit removes the automatic incrementing `self._id` property on KML elements,
so that the resulting KML output does not get littered with `id="..."` attributes.

    # OLD
    <Document id="1"><Placemark id="3">...

    # NEW
    <Document><Placemark>

This auto-incrementing value is still applied to document elements within a new
`self._global_id` attribute, but currently this is not being used.  If someone
wanted to continue the prior behavior using a flag, it could be enabled within
the `id` property method.

Furthermore, the `id` attribute can now be set during element construction to
allow for more definitive KML documents to be generated.

    kml = simplekml.Kml(id="sample")
    kml.newpoint(id="sample_point", coords=[[30,-30]])
    kml.kml(format=False)

    # <kml ...>
    #   <Document id="sample">
    #     <Placemark id="sample_point">
    #       <Point><coordinates>30,-30,0.0</coordinates></Point>
    #     </Placemark>
    #   </Document>
    # </kml>

Some of the `__str__()` methods were standardized during the refactor.