Changed coordinates to only include altitude if provided in input coordinates
The KML standard says that the altitude (m) coordinates are optional, so it is misleading
to default to "0.0" when they are omitted from the input feature.

This commit refactors the `Coordinates` object to avoid implicitly setting `0.0`
for the the altitude element, and simplifies the serialization of the `lng,lat[,alt]`
output using Python list comprehensions.